moved typing indicator to a separate component
