+++ src/components/character-typing/index.ts
@@ -1,0 +1,1 @@
+export * from "./character-typing-indicator.js";
